// parcial2023/src/app/models/datos.ts
export interface Datos
{
    ID: number;
    FirstName: string;
    LastName: string;
    Adress: string;
    State: string;
    City: string;
    IdentificationTipe: string;
    IdentificationNumber: string;

    NameCard: string;
    CreditCardNumber: string;
    CVV: string;
    ExpirationDate: string;
}